Update README.md
"John Wick, Oppenheimer, and The Roundup hold a special place among my favorite movies due to the rich reservoir of life lessons they offer. Each film, in its own unique way, provides a narrative tapestry that extends beyond mere entertainment, presenting valuable insights and reflections on the complexities of life. Whether it's the intense action sequences in John Wick, the historical depth of Oppenheimer, or the intriguing narrative of The Roundup, each cinematic experience has left an indelible mark on me, offering not only entertainment but also a profound exploration of the human experience."